+++ packages/twenty-ui/src/display/callout/Callout.tsx
@@ -1,52 +1,53 @@
+import { IconHelp, IconX } from 'twenty-ui/display';
 import styled from '@emotion/styled';
-import { IconInfoCircle, IconX } from 'twenty-ui/display';
 import { IconButton } from 'twenty-ui/input';
 
 type CalloutVariant = 'info' | 'warning' | 'error' | 'neutral' | 'success';
 
-const variantConfig = {
-  info: {
-    iconColor: '#3498db',
-    borderColor: '#3498db',
-    backgroundColor: '#f0f8ff',
-  },
-  warning: {
-    iconColor: '#ff9500',
-    borderColor: '#ff9500',
-    backgroundColor: '#fff8f0',
-  },
-  error: {
-    iconColor: '#ef4444',
-    borderColor: '#ef4444',
-    backgroundColor: '#fef2f2',
-  },
-  neutral: {
-    iconColor: '#9ca3af',
-    borderColor: '#e5e7eb',
-    backgroundColor: '#f9fafb',
-  },
-  success: {
-    iconColor: '#ffc107',
-    borderColor: '#ffc107',
-    backgroundColor: '#ffc107',
-  }
-};
-
 const StyledCalloutContainer = styled.div<{ variant: CalloutVariant }>`
-    background-color: ${({ variant }) => variantConfig[variant].backgroundColor};
-    border: 1px solid ${({ variant }) => variantConfig[variant].borderColor};
-    border-radius: 1px;
-    box-sizing: border-box;
-    display: flex;
-    gap: ${({ theme }) => theme.spacing(3)};
-    padding: ${({ theme }) => theme.spacing(4)};
-    position: relative;
-    width: 100%;
+  background-color: ${({ theme, variant }) =>
+  variant === 'info'
+    ? theme.color.blue1
+    : variant === 'warning'
+      ? theme.color.yellow1
+      : variant === 'success'
+        ? theme.color.green1
+        : variant === 'error'
+          ? theme.color.red1
+          : theme.color.gray1};
+  border: 1px solid
+    ${({ theme, variant }) =>
+  variant === 'info'
+    ? theme.color.blue6
+    : variant === 'warning'
+      ? theme.color.yellow6
+      : variant === 'success'
+        ? theme.color.green6
+        : variant === 'error'
+          ? theme.color.red6
+          : theme.color.gray6};
+  border-radius: 8px;
+  box-sizing: border-box;
+  display: flex;
+  gap: ${({ theme }) => theme.spacing(2)};
+  padding: ${({ theme }) => theme.spacing(4)};
+  position: relative;
+    margin-bottom: ${({ theme }) => theme.spacing(2)};
+    margin-left: ${({ theme }) => theme.spacing(7)};
 `;
 
 const StyledIconContainer = styled.div<{ variant: CalloutVariant }>`
   align-items: flex-start;
-  color: ${({ variant }) => variantConfig[variant].iconColor};
+  color: ${({ theme, variant }) =>
+  variant === 'info'
+    ? theme.color.blue9
+    : variant === 'warning'
+      ? theme.color.orange9
+      : variant === 'success'
+        ? theme.color.green9
+        : variant === 'error'
+          ? theme.color.red9
+          : theme.color.gray9};
   display: flex;
   flex-shrink: 0;
   padding-top: ${({ theme }) => theme.spacing(0.5)};
@@ -69,11 +70,11 @@ const StyledTitle = styled.div`
 `;
 
 const StyledDescription = styled.div`
-  color: ${({ theme }) => theme.font.color.secondary};
+  color: ${({ theme }) => theme.font.color.tertiary};
   font-family: ${({ theme }) => theme.font.family};
   font-size: ${({ theme }) => theme.font.size.sm};
   font-weight: ${({ theme }) => theme.font.weight.regular};
-  line-height: 1.6;
+  line-height: 15px;
 `;
 
 const StyledFooter = styled.div`
@@ -84,10 +85,10 @@ const StyledFooter = styled.div`
 `;
 
 const StyledLearnMoreLink = styled.a`
-  color: ${({ theme }) => theme.font.color.tertiary};
+  color: ${({ theme }) => theme.font.color.secondary};
   cursor: pointer;
   font-family: ${({ theme }) => theme.font.family};
-  font-size: ${({ theme }) => theme.font.size.sm};
+  font-size: ${({ theme }) => theme.font.size.md};
   font-weight: ${({ theme }) => theme.font.weight.regular};
   text-decoration: none;
   transition: color ${({ theme }) => theme.animation.duration.instant}s ease;
@@ -104,51 +105,51 @@ const StyledCloseButton = styled(IconButton)`
   top: ${({ theme }) => theme.spacing(3)};
 `;
 
-export type WorkflowFormCalloutProps = {
+export type CalloutProps = {
   variant: CalloutVariant;
   title: string;
   description: string;
+  learnMoreText: string;
   learnMoreUrl: string;
   onClose: () => void;
   className: string;
 };
 
 export const Callout = ({
-                                      variant,
-                                      title,
-                                      description,
-                                      learnMoreUrl,
-                                      onClose,
-                                      className,
-                                    }: WorkflowFormCalloutProps) => {
-
+                          variant,
+                          title,
+                          description,
+                          learnMoreText,
+                          learnMoreUrl,
+                          onClose,
+                          className,
+                        }: CalloutProps) => {
   return (
     <StyledCalloutContainer variant={variant} className={className}>
       <StyledIconContainer variant={variant}>
-        <IconInfoCircle size={16} />
+        <IconHelp size={16} />
       </StyledIconContainer>
 
       <StyledContent>
         <StyledTitle>{title}</StyledTitle>
         <StyledDescription>{description}</StyledDescription>
-          <StyledFooter>
-            <StyledLearnMoreLink
-              href={learnMoreUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Learn more
-            </StyledLearnMoreLink>
-          </StyledFooter>
+        <StyledFooter>
+          <StyledLearnMoreLink
+            href={learnMoreUrl}
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            {learnMoreText}
+          </StyledLearnMoreLink>
+        </StyledFooter>
       </StyledContent>
-
-        <StyledCloseButton
-          Icon={IconX}
-          size="small"
-          variant="tertiary"
-          onClick={onClose}
-          ariaLabel={`Close`}
-        />
+      <StyledCloseButton
+        Icon={IconX}
+        size="small"
+        variant="tertiary"
+        ariaLabel={`Close`}
+        onClick={onClose}
+      />
     </StyledCalloutContainer>
   );
 };
