+++ packages/twenty-ui/src/display/index.ts
@@ -19,7 +19,7 @@ export type { BannerVariant } from './banner/components/Banner';
 export { Banner } from './banner/components/Banner';
 export type { SidePanelInformationBannerProps } from './banner/components/SidePanelInformationBanner';
 export { SidePanelInformationBanner } from './banner/components/SidePanelInformationBanner';
-export type { WorkflowFormCalloutProps } from './callout/Callout';
+export type { CalloutProps } from './callout/Callout';
 export { Callout } from './callout/Callout';
 export type { AnimatedCheckmarkProps } from './checkmark/components/AnimatedCheckmark';
 export { AnimatedCheckmark } from './checkmark/components/AnimatedCheckmark';
